Before:
BenchmarkSimulateLootRNG-4                32          42091890 ns/op
PASS
ok      github.com/therizk/benchmarking 3.189s

After:
BenchmarkSimulateLootRNG-4            45          24992548 ns/op
PASS
ok      github.com/therizk/benchmarking 1.155s